fix: [DEL-4288]: Revamp NG secret manager (#34246) (#34479)
* fix: [DEL-4288]: Revamp NG secret manager (#34246)

Summary: Modified the secret cache in CG manager, to store encryption details as value. This will help us in determining which secret needs to be decrypted locally and which is not. In case of NG secrets, Rest api call used to happen to fetch these details to CG manager, which is expensive. And now since we are caching these details in CG manager itself, these latencies will go away.

* Bumped up build number

* fix: [DEL-4288]: Changed cache name