fix:[CDS-58178]: Skip audit operation when the execution comes from an API call
The audit operation of an API call is made from AuthRuleFilter without identify which user made the call, then the service layer don't need to audit it again as we don't know which user made it. That is a CG design as the API token is related to a group of users instead of one user.