fix:[CDS-58178]: Don't add user details when header comes from API call (#46133)
The audit operation of an API call is made from AuthRuleFilter without identifying which user made the call, then the service layer doesn't need to audit it again as we don't know which user made it. That is a CG design as the API token is related to a group of users instead of one user.